Rework the demo a little bit to use an array to hold the geojson instead of an object. Adding a collate_geometry.js tiny helper function to recursively collate all fo the geojson in the data folder into a single array with normalized labels and options-names
