--- data/edgecase_geo/anua.json
@@ -1,7 +1,0 @@
-{
-"type": "FeatureCollection",
-"crs": { "type": "name", "properties": { "name": "urn:ogc:def:crs:EPSG::4269" } },
-"features": [
-{ "type": "Feature", "properties": { "STATEFP": "60", "PLACEFP": "11300", "PLACENS": "02414199", "GEOID": "6011300", "NAME": "test_Anua", "NAMELSAD": "Anua village", "LSAD": "47", "CLASSFP": "C1", "PCICBSA": "N", "PCINECTA": "N", "MTFCC": "G4110", "FUNCSTAT": "A", "ALAND": 180622, "AWATER": 130582, "INTPTLAT": "-14.2710270", "INTPTLON": "-170.6873951", "State": null, "County": null, "City": null, "name_2": null, "RegionID": null, "join_STATE": null, "join_PLACE": null, "join_PLA_1": null, "join_GEOID": null, "join_NAME": null, "join_NAMEL": null, "join_LSAD": null, "join_CLASS": null, "join_PCICB": null, "join_PCINE": null, "join_MTFCC": null, "join_FUNCS": null, "join_ALAND": null, "join_AWATE": null, "join_INTPT": null, "join_INT_1": null, "distance": null, "STATEFP_1": "60", "COUNTYFP": "010", "COUNTYNS": "01805240", "AFFGEOID": "0500000US60010", "GEOID_1": "60010", "NAME_1": "Eastern", "LSAD_1": "07", "ALAND_1": 65124867, "AWATER_1": 349724676 }, "geometry": { "type": "MultiPolygon", "coordinates": [ [ [ [ -170.685593512638007, -14.271070063115365 ], [ -170.685600135293015, -14.271080869487299 ], [ -170.685838, -14.271469 ], [ -170.689252213955996, -14.271821857896599 ], [ -170.690263040945013, -14.271926326535574 ], [ -170.688926818535691, -14.266875044160555 ], [ -170.685593512638007, -14.271070063115365 ] ] ], [ [ [ -170.68802006627368, -14.277038452894571 ], [ -170.687359, -14.276398 ], [ -170.685175, -14.276571 ], [ -170.682856, -14.276042 ], [ -170.682845, -14.27604 ], [ -170.682797278778509, -14.276044908468494 ], [ -170.686055916016102, -14.277835975883914 ], [ -170.68802006627368, -14.277038452894571 ] ] ] ] } }
-]
-}
