+++ data/edgecase_geo/cape_may_nj.json
@@ -1,1 +1,1 @@
-{ "type": "Feature", "properties": { "STATEFP": "34", "COUNTYFP": "009", "COUNTYNS": "00882274", "AFFGEOID": "0500000US34009", "GEOID": "34009", "NAME": "test_Cape May", "LSAD": "06", "ALAND": 651209688, "AWATER": 955666476 }, "geometry": { "type": "Polygon", "coordinates": [ [ [ -74.955363, 39.001262 ], [ -74.903664, 39.087437 ], [ -74.885914, 39.143627 ], [ -74.905181, 39.174945 ], [ -74.915155805562705, 39.176697503518795 ], [ -74.925952, 39.19714 ], [ -74.861073, 39.323768 ], [ -74.65259, 39.288566 ], [ -74.540985007062687, 39.299882326532597 ], [ -74.581008, 39.270819 ], [ -74.646595, 39.212002 ], [ -74.67143, 39.179802 ], [ -74.714341, 39.119804 ], [ -74.705876, 39.102937 ], [ -74.778777, 39.023073 ], [ -74.807917, 38.985948 ], [ -74.864458, 38.94041 ], [ -74.933571, 38.928519 ], [ -74.967274, 38.933413 ], [ -74.955363, 39.001262 ] ] ] } }
+{ "type": "Feature", "properties": { "STATEFP": "34", "COUNTYFP": "009", "COUNTYNS": "00882274", "AFFGEOID": "0500000US34009", "GEOID": "34009", "NAME": "Cape May", "LSAD": "06", "ALAND": 651209688, "AWATER": 955666476 }, "geometry": { "type": "Polygon", "coordinates": [ [ [ -74.955363, 39.001262 ], [ -74.903664, 39.087437 ], [ -74.885914, 39.143627 ], [ -74.905181, 39.174945 ], [ -74.915155805562705, 39.176697503518795 ], [ -74.925952, 39.19714 ], [ -74.861073, 39.323768 ], [ -74.65259, 39.288566 ], [ -74.540985007062687, 39.299882326532597 ], [ -74.581008, 39.270819 ], [ -74.646595, 39.212002 ], [ -74.67143, 39.179802 ], [ -74.714341, 39.119804 ], [ -74.705876, 39.102937 ], [ -74.778777, 39.023073 ], [ -74.807917, 38.985948 ], [ -74.864458, 38.94041 ], [ -74.933571, 38.928519 ], [ -74.967274, 38.933413 ], [ -74.955363, 39.001262 ] ] ] } }
