+++ data/edgecase_geo/childress_tx.json
@@ -1,1 +1,1 @@
-{ "type": "Feature", "properties": { "STATEFP": "48", "COUNTYFP": "075", "COUNTYNS": "01383823", "AFFGEOID": "0500000US48075", "GEOID": "48075", "NAME": "test_Childress", "LSAD": "06", "ALAND": 1803680231, "AWATER": 44835076 }, "geometry": { "type": "Polygon", "coordinates": [ [ [ -100.416839, 34.42457 ], [ -100.415864, 34.558013 ], [ -100.415786, 34.558506 ], [ -100.415794, 34.560651 ], [ -100.41585, 34.620963 ], [ -100.415895, 34.747525 ], [ -100.382295, 34.747524 ], [ -100.375398, 34.747128 ], [ -100.274939, 34.74713 ], [ -100.000812, 34.746444 ], [ -100.000381, 34.746358 ], [ -100.000381, 34.616297 ], [ -100.000381, 34.570426199215696 ], [ -100.000381, 34.560509 ], [ -99.997501410946697, 34.560423887043399 ], [ -99.997831, 34.463389 ], [ -99.997816, 34.448965 ], [ -99.99784, 34.441912 ], [ -99.997812, 34.400871 ], [ -99.998006, 34.396821 ], [ -99.997832, 34.390816 ], [ -99.99772, 34.311829 ], [ -100.000404, 34.311929 ], [ -100.024285, 34.312002 ], [ -100.12318, 34.31229 ], [ -100.134601, 34.311976 ], [ -100.24152, 34.312229 ], [ -100.350099, 34.312713 ], [ -100.350774, 34.312571 ], [ -100.359546, 34.312561 ], [ -100.375394, 34.312718 ], [ -100.399581, 34.31354 ], [ -100.417783, 34.313524 ], [ -100.416839, 34.42457 ] ] ] } }
+{ "type": "Feature", "properties": { "STATEFP": "48", "COUNTYFP": "075", "COUNTYNS": "01383823", "AFFGEOID": "0500000US48075", "GEOID": "48075", "NAME": "Childress", "LSAD": "06", "ALAND": 1803680231, "AWATER": 44835076 }, "geometry": { "type": "Polygon", "coordinates": [ [ [ -100.416839, 34.42457 ], [ -100.415864, 34.558013 ], [ -100.415786, 34.558506 ], [ -100.415794, 34.560651 ], [ -100.41585, 34.620963 ], [ -100.415895, 34.747525 ], [ -100.382295, 34.747524 ], [ -100.375398, 34.747128 ], [ -100.274939, 34.74713 ], [ -100.000812, 34.746444 ], [ -100.000381, 34.746358 ], [ -100.000381, 34.616297 ], [ -100.000381, 34.570426199215696 ], [ -100.000381, 34.560509 ], [ -99.997501410946697, 34.560423887043399 ], [ -99.997831, 34.463389 ], [ -99.997816, 34.448965 ], [ -99.99784, 34.441912 ], [ -99.997812, 34.400871 ], [ -99.998006, 34.396821 ], [ -99.997832, 34.390816 ], [ -99.99772, 34.311829 ], [ -100.000404, 34.311929 ], [ -100.024285, 34.312002 ], [ -100.12318, 34.31229 ], [ -100.134601, 34.311976 ], [ -100.24152, 34.312229 ], [ -100.350099, 34.312713 ], [ -100.350774, 34.312571 ], [ -100.359546, 34.312561 ], [ -100.375394, 34.312718 ], [ -100.399581, 34.31354 ], [ -100.417783, 34.313524 ], [ -100.416839, 34.42457 ] ] ] } }
