+++ data/edgecase_geo/index.js
@@ -1,10 +1,8 @@
 module.exports = [
-  require('./aleutians_west_al.json'),
   require('./cape_may_nj.json'),
   require('./childress_tx.json'),
   require('./reagan_tx.json'),
   require('./roanoke_va.json'),
   require('./wise_va.json'),
-  require('./anua.json'),
   require('./dukes_ma.json')
 ]
