+++ data/edgecase_geo/reagan_tx.json
@@ -1,1 +1,1 @@
-{ "type": "Feature", "properties": { "STATEFP": "48", "COUNTYFP": "383", "COUNTYNS": "01383977", "AFFGEOID": "0500000US48383", "GEOID": "48383", "NAME": "test_Reagan", "LSAD": "06", "ALAND": 3044049124, "AWATER": 1792716 }, "geometry": { "type": "Polygon", "coordinates": [ [ [ -101.775808, 31.122899 ], [ -101.775544, 31.125154 ], [ -101.775381, 31.36064 ], [ -101.775654, 31.363379 ], [ -101.775802, 31.651319 ], [ -101.500396, 31.651451 ], [ -101.498368, 31.651336 ], [ -101.375394, 31.651091 ], [ -101.369115, 31.651265 ], [ -101.267122, 31.650854 ], [ -101.266981, 31.644192 ], [ -101.26763, 31.556462 ], [ -101.267947, 31.528688 ], [ -101.268452, 31.48192 ], [ -101.268617, 31.478798 ], [ -101.268473, 31.474473 ], [ -101.270726, 31.347593 ], [ -101.271888, 31.294032 ], [ -101.274799, 31.079379 ], [ -101.375393, 31.079485 ], [ -101.456382, 31.079302 ], [ -101.776191, 31.079784 ], [ -101.775808, 31.122899 ] ] ] } }
+{ "type": "Feature", "properties": { "STATEFP": "48", "COUNTYFP": "383", "COUNTYNS": "01383977", "AFFGEOID": "0500000US48383", "GEOID": "48383", "NAME": "Reagan", "LSAD": "06", "ALAND": 3044049124, "AWATER": 1792716 }, "geometry": { "type": "Polygon", "coordinates": [ [ [ -101.775808, 31.122899 ], [ -101.775544, 31.125154 ], [ -101.775381, 31.36064 ], [ -101.775654, 31.363379 ], [ -101.775802, 31.651319 ], [ -101.500396, 31.651451 ], [ -101.498368, 31.651336 ], [ -101.375394, 31.651091 ], [ -101.369115, 31.651265 ], [ -101.267122, 31.650854 ], [ -101.266981, 31.644192 ], [ -101.26763, 31.556462 ], [ -101.267947, 31.528688 ], [ -101.268452, 31.48192 ], [ -101.268617, 31.478798 ], [ -101.268473, 31.474473 ], [ -101.270726, 31.347593 ], [ -101.271888, 31.294032 ], [ -101.274799, 31.079379 ], [ -101.375393, 31.079485 ], [ -101.456382, 31.079302 ], [ -101.776191, 31.079784 ], [ -101.775808, 31.122899 ] ] ] } }
